Eliminate "useless cast" warnings related to AV_NOPTS_VALUE.
Running GCC with the "useless-cast" warning enabled produces a number
of warnings.  The definition of AV_NOPTS_VALUE includes a cast to
int64_t, so there is no need to cast it when the definition is used.

+++ mythplugins/mytharchive/mytharchive/thumbfinder.cpp
@@ -805,7 +805,7 @@ bool ThumbFinder::getFrameImage(bool needKeyFrame, int64_t requiredPTS)
 
             keyFrame = pkt.flags & AV_PKT_FLAG_KEY;
 
-            if (m_startPTS == -1 && pkt.dts != (int64_t)AV_NOPTS_VALUE)
+            if (m_startPTS == -1 && pkt.dts != AV_NOPTS_VALUE)
             {
                 m_startPTS = pkt.dts;
                 m_frameTime = pkt.duration;
@@ -830,7 +830,7 @@ bool ThumbFinder::getFrameImage(bool needKeyFrame, int64_t requiredPTS)
                 frameFinished = 1;
             if (ret == 0 || ret == AVERROR(EAGAIN))
                 ret = avcodec_send_packet(m_codecCtx, &pkt);
-            if (requiredPTS != -1 && pkt.dts != (int64_t)AV_NOPTS_VALUE && pkt.dts < requiredPTS)
+            if (requiredPTS != -1 && pkt.dts != AV_NOPTS_VALUE && pkt.dts < requiredPTS)
                 frameFinished = false;
 
             m_currentPTS = pkt.dts;
